pyproject.toml
[project]
name = "livekit-agents-examples-ja"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "flask>=3.1.0",
    "livekit-agents>=1.0.0rc9",
    "livekit-plugins-anthropic>=1.0.0rc9",
    "livekit-plugins-cartesia>=1.0.0rc9",
    "livekit-plugins-deepgram>=1.0.0rc9",
    "livekit-plugins-elevenlabs>=1.0.0rc9",
    "livekit-plugins-google>=1.0.0rc9",
    "livekit-plugins-groq>=1.0.0rc9",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.0.0rc9",
    "livekit-plugins-playai>=1.0.0rc9",
    "livekit-plugins-rime>=1.0.0rc9",
    "livekit-plugins-silero>=1.0.0rc9",
    "livekit-plugins-turn-detector>=1.0.0rc9",
    "mcp>=1.6.0",
    "onnxruntime==1.20.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "sounddevice>=0.5.1",
    "websockets>=11.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.5",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["livekit"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
strict = true
ignore_missing_imports = true
